+++ src/Common/src/TypeSystem/IL/Stubs/PInvokeILEmitter.cs
@@ -283,77 +283,6 @@ namespace Internal.IL.Stubs
             UnmarshallingCodestream = Emitter.NewCodeStream();
         }
     }
-    /// <summary>
-    /// Synthetic method that represents the actual PInvoke target method.
-    /// All parameters are simple types. There will be no code
-    /// generated for this method. Instead, a static reference to a symbol will be emitted.
-    /// </summary>
-    public sealed partial class PInvokeTargetNativeMethod : MethodDesc
-    {
-        private readonly MethodDesc _declMethod;
-        private readonly MethodSignature _signature;
-        
-        public PInvokeTargetNativeMethod(MethodDesc declMethod, MethodSignature signature)
-        {
-            _declMethod = declMethod;
-            _signature = signature;
-        }
-
-        public override TypeSystemContext Context
-        {
-            get
-            {
-                return _declMethod.Context;
-            }
-        }
-
-        public override TypeDesc OwningType
-        {
-            get
-            {
-                return _declMethod.OwningType;
-            }
-        }
-
-        public override MethodSignature Signature
-        {
-            get
-            {
-                return _signature;
-            }
-        }
-
-        public override string Name
-        {
-            get
-            {
-                return _declMethod.Name;
-            }
-        }
-
-        public override bool HasCustomAttribute(string attributeNamespace, string attributeName)
-        {
-            return false;
-        }
-
-        public override bool IsPInvoke
-        {
-            get
-            {
-                return true;
-            }
-        }
-
-        public override PInvokeMetadata GetPInvokeMethodMetadata()
-        {
-            return _declMethod.GetPInvokeMethodMetadata();
-        }
-
-        public override string ToString()
-        {
-            return "[EXTERNAL]" + Name;
-        }
-    }
 
     /// <summary>
     /// Synthetic RVA static field that represents PInvoke fixup cell. The RVA data is
